renamed gateway server into manager
